[Fix] : #74 CodeRabbit 타당성 반영 — TopSku.totalQuantity long(집계 오버플로 방지) + top-SKU 동률 시 sku 2차키로 결정성 보장

+++ src/main/java/com/bbd/sales/adapter/in/web/dto/SalesOrderStatsResponse.java
@@ -11,6 +11,6 @@ public record SalesOrderStatsResponse(
     public record Backorder(long count, double avgWaitDays, long maxWaitDays, List<TopSku> topSkus) {
     }
 
-    public record TopSku(String sku, String name, long lineCount, int totalQuantity) {
+    public record TopSku(String sku, String name, long lineCount, long totalQuantity) {
     }
 }
